feat: 방/통계 카드 재사용 고려 (#87) - 방 카드 렌더링을 재사용 가능하도록 RoomCardList 컴포넌트로 분리 - 통계 카드(진행 중 배틀/총 관전자/참가 플레이어)를 데이터 배열로 map 렌더링하도록 단순화
