+++ apps/web/src/pages/MainPage.tsx
@@ -1,9 +1,9 @@
 import type { Room } from '@shared/types/room';
-import { Eye } from 'lucide-react';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import Header from '@/components/Header/Header';
+import RoomCardList from '@/components/Main/RoomCardList';
 import { useBattleSocketStore } from '@/stores/battleSocketStore';
 import { useMatchingStore } from '@/stores/matchingStore';
 import { useUserStore } from '@/stores/userStore';
@@ -33,7 +33,7 @@ function MainPage() {
   const startMatching = useMatchingStore((state) => state.startMatching);
   const registerMatchingListeners = useMatchingStore((state) => state.registerMatchingListeners);
 
-  const [joiningRoomId] = useState<string | null>(null);
+  //const [joiningRoomId] = useState<string | null>(null);
 
   useEffect(() => {
     const socket = connect();
@@ -112,9 +112,6 @@ function MainPage() {
   };
 
   // 관전하기 입장 로직은 추후 백엔드 연동 시 추가 예정
-  const handleJoinSpectator = () => {};
-
-  const getInitial = (name?: string) => name?.trim().charAt(0)?.toUpperCase() ?? '?';
 
   return (
     <div className="min-h-screen bg-bg-layer-1">
@@ -157,93 +154,23 @@ function MainPage() {
 
         {/* 통계 카드 */}
         <div className="grid grid-cols-1 gap-4 md:grid-cols-3">
-          <div className="rounded-2xl bg-white px-6 py-4 shadow-sm">
-            <p className="text-sm font-semibold text-green-05">진행 중인 배틀</p>
-            <p className="mt-2 text-3xl font-bold text-green-05">{stats.totalBattles}</p>
-          </div>
-          <div className="rounded-2xl bg-white px-6 py-4 shadow-sm">
-            <p className="text-sm font-semibold text-green-05">총 관전자</p>
-            <p className="mt-2 text-3xl font-bold text-green-05">{stats.totalSpectators}</p>
-          </div>
-          <div className="rounded-2xl bg-white px-6 py-4 shadow-sm">
-            <p className="text-sm font-semibold text-green-05">참가 중인 플레이어</p>
-            <p className="mt-2 text-3xl font-bold text-green-05">{stats.totalPlayers}</p>
-          </div>
+          {[
+            { label: '진행 중인 배틀', value: stats.totalBattles },
+            { label: '총 관전자', value: stats.totalSpectators },
+            { label: '참가 중인 플레이어', value: stats.totalPlayers },
+          ].map((item) => (
+            <div
+              key={item.label}
+              className="rounded-2xl bg-bg-layer-2 border border-border-soft px-6 py-4 shadow-sm"
+            >
+              <p className="text-3xl font-bold text-green-05">{item.value}</p>
+              <p className="mt-1 text-sm text-base-secondary">{item.label}</p>
+            </div>
+          ))}
         </div>
 
         {/* 방 카드 리스트 */}
-        <div className="grid grid-cols-1 gap-6 lg:grid-cols-2">
-          {roomCards.map((room) => {
-            const players = room.currentPlayers?.slice(0, 2) ?? [];
-            return (
-              <div
-                key={room.roomId}
-                className="rounded-2xl bg-white px-5 py-4 shadow-sm transition hover:shadow-md"
-              >
-                <div className="mb-4 flex items-center justify-between text-xs text-base-secondary">
-                  <div className="flex items-center gap-2">
-                    <span className="flex items-center font-semibold text-red-01">● LIVE</span>
-                    <span className="text-blue-03">12:34</span>
-                  </div>
-                  <div className="inline-flex items-center gap-1.5 text-blue-03 text-sm font-medium">
-                    <Eye className="h-4 w-4 text-blue-03" strokeWidth={1.5} />
-                    {room.currentSpectators?.length ?? 0}
-                  </div>
-                </div>
-                <div className="flex items-center gap-2">
-                  <h3 className="text-xl font-bold text-base-primary">{room.title}</h3>
-                  <span className="rounded-full bg-green-01 px-2 py-1 text-[10px] font-bold text-green-06">
-                    Gold
-                  </span>
-                </div>
-
-                <div className="mt-4 space-y-4">
-                  {players.map((p, idx) => (
-                    <div key={p.userId} className="space-y-1">
-                      <div className="flex items-center gap-3">
-                        <div className="flex h-9 w-9 items-center justify-center rounded-full bg-base-muted text-sm font-bold text-base-primary">
-                          {p.avatarUrl ? (
-                            <img
-                              src={p.avatarUrl}
-                              alt={p.username}
-                              className="h-full w-full rounded-full object-cover"
-                            />
-                          ) : (
-                            getInitial(p.username)
-                          )}
-                        </div>
-                        <div className="flex flex-col">
-                          <span className="text-sm font-semibold text-base-primary">
-                            {p.username}
-                          </span>
-                          <span className="text-xs text-base-secondary">🏆 Gold III</span>
-                        </div>
-                      </div>
-                      <div className="h-2 w-full overflow-hidden rounded-full bg-base-muted">
-                        <div
-                          className="h-full rounded-full"
-                          style={{
-                            width: `${idx === 0 ? 75 : 60}%`,
-                            backgroundColor: idx === 0 ? 'bg-green-05' : 'bg-red-01',
-                          }}
-                        />
-                      </div>
-                    </div>
-                  ))}
-                </div>
-
-                <button
-                  type="button"
-                  onClick={() => handleJoinSpectator()}
-                  className="mt-6 flex w-full items-center justify-center rounded-lg bg-base-muted px-4 py-3 text-sm font-semibold text-base-primary transition hover:brightness-105 disabled:opacity-50"
-                  disabled={joiningRoomId === room.roomId}
-                >
-                  관전하기
-                </button>
-              </div>
-            );
-          })}
-        </div>
+        <RoomCardList rooms={roomCards} onSpectate={undefined} />
       </main>
     </div>
   );
